refactor: reorganize UserSetting-related packages from `entity` to `model` and update imports
