+++ src/main/java/com/yeoljeong/tripmate/usersetting/application/UserSettingQueryService.java
@@ -4,7 +4,7 @@ import static com.yeoljeong.tripmate.usersetting.domain.exception.UserSettingErr
 
 import com.yeoljeong.tripmate.exception.BusinessException;
 import com.yeoljeong.tripmate.usersetting.application.dto.result.UserSettingResult;
-import com.yeoljeong.tripmate.usersetting.domain.entity.UserSetting;
+import com.yeoljeong.tripmate.usersetting.domain.model.UserSetting;
 import com.yeoljeong.tripmate.usersetting.domain.repository.UserSettingRepository;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
