--- src/main/java/com/yeoljeong/tripmate/usersetting/domain/entity/constants/MbtiIE.java
@@ -1,5 +1,0 @@
-package com.yeoljeong.tripmate.usersetting.domain.entity.constants;
-
-public enum MbtiIE {
-	I, E, UNKNOWN
-}
