--- src/main/java/com/yeoljeong/tripmate/usersetting/domain/entity/constants/MbtiPJ.java
@@ -1,5 +1,0 @@
-package com.yeoljeong.tripmate.usersetting.domain.entity.constants;
-
-public enum MbtiPJ {
-	P, J, UNKNOWN
-}
