+++ src/main/java/com/yeoljeong/tripmate/usersetting/domain/model/constants/Gender.java
@@ -1,0 +1,5 @@
+package com.yeoljeong.tripmate.usersetting.domain.model.constants;
+
+public enum Gender {
+	MALE, FEMALE
+}
