+++ src/main/java/com/yeoljeong/tripmate/usersetting/domain/repository/UserSettingRepository.java
@@ -1,6 +1,6 @@
 package com.yeoljeong.tripmate.usersetting.domain.repository;
 
-import com.yeoljeong.tripmate.usersetting.domain.entity.UserSetting;
+import com.yeoljeong.tripmate.usersetting.domain.model.UserSetting;
 import java.util.Optional;
 import java.util.UUID;
