+++ src/main/java/com/yeoljeong/tripmate/usersetting/infrastructure/persistence/repositoryimpl/UserSettingJpaRepository.java
@@ -1,6 +1,6 @@
 package com.yeoljeong.tripmate.usersetting.infrastructure.persistence.repositoryimpl;
 
-import com.yeoljeong.tripmate.usersetting.domain.entity.UserSetting;
+import com.yeoljeong.tripmate.usersetting.domain.model.UserSetting;
 import com.yeoljeong.tripmate.usersetting.domain.repository.UserSettingRepository;
 import com.yeoljeong.tripmate.usersetting.infrastructure.persistence.jpa.SpringDataUserSettingRepository;
 import java.util.Optional;
